gluster: Publish gluster get volume info
In 3.6 we use glusterVolumeInfo API to get gluster volume info in order to
mount gluster to backup volume servers.  glusterVolumeInfo API is defined in
vdsm/gluster/cli.py and is imported to supervdsm. In vdsm/gluster/cli.py there
are APIs that are used to manage gluster volumes; however, we don't want to
expose all these APIs in supervdsm unless gluster is used, while we do want to
expose getGlusterVolumeInfo.  To achieve this goal, we expose RHEV_ENABLED
param. With this change, gluster API defined in vdsm/gluster/cli.py exposed on
supervdsm either when building with gluster or when building on RHEV.

Change-Id: Ib9c70e9afa2fd776dfd9c2767c2e83e7f179b5c2
Bug-Url: https://bugzilla.redhat.com/1254499
Signed-off-by: Ala Hino <ahino@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/45098
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Bala.FA <barumuga@redhat.com>
Continuous-Integration: Jenkins CI