networkTests: normalize strings (to unicode) through json magic
To make it easier to read differences between "running config"
and "kernel config" (the state of networking) the strings
contained in dictionaries must not be stored randomly as str
or unicode instances.

This patch converts all strings to unicode. Thanks to Ido Barkan
for a neat trick using the json library.

Change-Id: I6362e04189f134ad27ee4e3c1991020b2eee91aa
Signed-off-by: Ondřej Svoboda <osvoboda@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/45201
Continuous-Integration: Jenkins CI
Reviewed-by: Ido Barkan <ibarkan@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>