caps: more precise emulated machines selection
Current VDSM code fetches emulated machines from the section which matches
first the required architecture.
This works in the simplest (and the only recommended) case, but doesn't
if, for any reason, there is more than one valid emulatore for a given
architecture.

The only known case is if the user managed to install qemu-kvm alongside
plain qemu.

This patch implements more robust capabilities fetching to deal with
this corner case.

Change-Id: I4deebbc90bf1cec53fc40bc6a35c6ada933296c3
Bug-Url: https://bugzilla.redhat.com/1239258
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/45257
Continuous-Integration: Jenkins CI
Reviewed-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/45850
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>