virt: vm with port mirroring  stuck in down state
When powering off VM with port mirroring enabled the
VM id down in the UI but stuck in down state in vdsm.

The problem was that the 'tc' command was invoked
twice when a vm was shutdown from the cli. The first
time as a result of a libvirt life cycle event which
deleted the qdisc entries for the vm, the second time
it was invoked by the engine and failed as the qdisc
entries were no longer there.
The code invoking tc was moved to be after the if
statement checking if the vm has already been released.

Change-Id: Ia43369f584f696f1387685ec8f4fc4b6f58d98a6
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1254713
Signed-off-by: Marcin Mirecki <mmirecki@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/45344
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/46394