make v2v errors reusable
We plan to introduce a generic jobs module which will be used for
monitoring and managing long running operations that are tied to a
specific host.  Errors for a job already existing (JobExistsError), job
not found (NoSuchJob), and job still running (JobNotDone) should be
usable by the new infrastructure.  Let's drop the V2V prefixes from the
names and update the error messages to remove v2v-specific language.

This does not affect the engine<-->vdsm API since the error codes remain
unchanged.

Change-Id: Ifa94b085d76a1e53eec7ec33839529d5a3b5962d
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/45381
Continuous-Integration: Jenkins CI
Reviewed-by: Shahar Havivi <shavivi@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>