Live Merge: Notice imageSyncVolumeChain errors
Currently we ignore errors from imageSyncVolumeChain when resolving a
completed live merge operation.  This could cause us to signal the
completion of a live merge operation to ovirt-engine without first
updating the storage metadata.  Instead, we should detect the error and
raise an exception so that the LiveMergeCleanupThread working on this
job will retry the operation later.

Change-Id: I5447d06b6aade97f71a3b7d6d7c73f9699c307d0
Bug-Url: https://bugzilla.redhat.com/1214342
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/45387
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Ala Hino <ahino@redhat.com>
Reviewed-by: Greg Padgett <gpadgett@redhat.com>
Tested-by: Greg Padgett <gpadgett@redhat.com>