StorageDomainManifest: Move acquireVolumeMetadataSlot
Moving acquireVolumeMetadataSlot into the manifest breaks volumeTests
which depended on mocking StorageDomain objects.  Since we already have
the manifest_tests infrastructure, it's better to move the test cases
implemented in volumeTests into manifest_tests and use a real
BlockStorageDomainManifest for testing.  In order to do this we need to
add an addtag utility method to our FakeLVM class.

Change-Id: I690ed4151428b22774415f38b350676391ccca46
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/45389
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>