netinfo: report DHCP on devices only when configured (like for networks)
As reporting of DHCP relies on reading of dhclient lease files, it is not
reliable when a network is reconfigured to static addressing: unused leases
indicate dhclient presence hours after it has been stopped (expiration time
is usually in hours or days).

The false indication has already been fixed for networks, now even their
respective devices adhere to configuration to stop false positives.

A year and half ago it was suggested first to rely on cmdlines on running
dhclients (where network device is the last parameter for all network
configurators). This fix is just a hacky, though little bandaid for the 3.6
release. A proper fix will be considered for the 4.0 branch.

testSetupNetworksAddDelDhcp now checks that DHCP stops being reported
on a network device, in addition to its network.

Change-Id: I2b17b0c66e3a0fc8c5ada62d166d9a2b92513033
Bug-Url: https://bugzilla.redhat.com/1184497
Signed-off-by: Ondřej Svoboda <osvoboda@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/45504
Continuous-Integration: Jenkins CI
Tested-by: Eliraz Levi <elevi@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/46247
Reviewed-by: Ido Barkan <ibarkan@redhat.com>