volume: fix failing metadata parsing
The call to "split" in file and block volumes expects only one "=" per
line. This is not a good behavior since there might be fields that
contain some more "=" (like Description).
This patch limits the number of splits to 1, so that the fields values
can contains some other "=".

Change-Id: Ie84bf5e42d2c4949944a5f348615d6fe80f72fa2
Backport-To: 3.6
Bug-Url: https://bugzilla.redhat.com/1258835
Signed-off-by: Idan Shaby <ishaby@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/45585
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Continuous-Integration: Jenkins CI