jsonrpc: fd leak
We need to make sure that there is no reference to objects which holds
a socket because m2c closes socket only when its object is garbage
collected.

This patch fixes two places where we had direct or indirect references
to socket object. In _StompConnection we kept a reference to it but now
we use dispatcher.socket. We passed connection object to JsonRpcServer
in order to get get local address which slowed down collection of socket
object (different thread) which could lead to reaching fd limit allowed
for the process. We removed this issue by passing local address to the
server instead.


Bug-Url: https://bugzilla.redhat.com/1256446
Change-Id: Ib15cca4659553b320babf928b603a0a75013ba99
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>
Reviewed-on: https://gerrit.ovirt.org/45714
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>