supervdsm: don't let _runAs return early due to EINTR
When _runAs executes a command, it polls a pipe to determine status of
that command.  The poll has a timeout, also allowing the child to be
killed if execution takes too long.  If a stray signal is received by
the parent process, it may cause the poll to return early, prematurely
ending child command execution ending in unintentional failure.

This is particularly prevalant when subsequent _runAs executions are
performed.  Termination of the first will result in a SIGCHLD which
interrupts polling for the second, as exposed by commit 777c36d6.

Because poll will return early without an exception if the low-level
call is interrupted, NoIntrCall() cannot be used and thus a new helper
function is provided that retries based on the time elapsed since the
call started.

PEP-475 will handle this in Python 3.5, but in the meantime this
workaround is needed.

Change-Id: I25af73a8fe67e7bc434f60a7d8492e33dc3b2ffa
Backport-To: 3.6
Bug-Url: https://bugzilla.redhat.com/1259310
Signed-off-by: Greg Padgett <gpadgett@redhat.com>
Reviewed-On: https://gerrit.ovirt.org/45752
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Tested-by: Greg Padgett <gdpadgett@gmail.com>
Reviewed-on: https://gerrit.ovirt.org/45885
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>