core: moving InquireNotSupportedError to storage_exception.py
InquireNotSupportedError is currently defined in clusterlock.py, that
prohibits from assigning a meaningful error code to that error and to
use it outside of that class scope without using a different method then
our wildly used one. In this patch its moved to storage_exception so
we'll be able catch and inspect that error like any other clusterlock related
error. The engine will use that error as well and will attempt to start
the spm if getSpmStatus() will fail as we don't have the "current" spm
information, on the worst case startSpm() will fail.

Change-Id: I8201794dc96ee24dc9c0da5b7c3d71ab0b75e9f3
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1269411
Signed-off-by: Liron Aravot <laravot@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/45975
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Continuous-Integration: Francesco Romani <fromani@redhat.com>