safelease: Improve logging when releasing lock
We used to log that "Cluster lock was released successfully" even if
spmstop.sh failed. Now we log either an error or a success message, but
never both.

To make spmstop.sh stderr and stdout more readable, we invoke spmstop.sh
with raw=True to avoid the unwanted line splitting.

We also log the domain UUID in all messages during release to make it
easier to follow the log.

Change-Id: Ica62c5281e19861f9bdcb0c2290627dc09370cd9
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/45961
Continuous-Integration: Jenkins CI
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>