utils.retry: Support system time changes while retry() is running
In a scenario where retry() is called with timeout (>0) and
during the retry run, the system time changes, the timeout calculation
will get corrupted. This may cause an unwanted result.

At the function entry point, the system time is saved (time.time())
and used after each retry attempt to calculate the period passed.
If the system time is changed between the first time call and the
following time calls, the delta time will get corrupted.

The calculated delta time, which gets compared to the timeout value,
should not be affected by the system time change.
To support that, a monotonic timer is used.

Change-Id: I6bc072bc20e8f5ae2db3fd1e48d91cdce9d469f8
Signed-off-by: Edward Haas <edwardh.dev@gmail.com>
Reviewed-on: https://gerrit.ovirt.org/46075
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Dan Kenigsberg <danken@redhat.com>