ifcfg: only determine and write HWADDR in removeNic if enabled
Writing of HWADDR has been deprecated for long now but for compatibility
with pre-1.0 NetworkManager (that matches on HWADDR in order to accept
NM_CONTROLLED=no) we shall still write it if requested with
hwaddr_in_ifcfg=always in vdsm.conf. This patch makes removeNic follow
suit -- addNic was alone in respecting the condition.

For more information refer to http://gerrit.ovirt.org/22288 and
https://bugzilla.redhat.com/1044060

Change-Id: Ieb4e7cdae83998b4102e79b7cb548ed95f672433
Bug-Url: https://bugzilla.redhat.com/1256252
Signed-off-by: Ondřej Svoboda <osvoboda@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/45983
Continuous-Integration: Jenkins CI
Reviewed-by: Ido Barkan <ibarkan@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>