vm: always hold confLock when updating pauseCode
An update of pauseCode can create a new field in Vm.conf.
If a client calls the getVMList() API, or for any other API
which needs to loop over all the VM conf, we can end up with

RuntimeError: dictionary changed size during iteration

This patch make sure that when we update Vm.conf with
pauseCode-related data, we always have Vm.confLock held.

Change-Id: Ie84e2442165fb524f2433fab7faa3b049b340038
Bug-Url: https://bugzilla.redhat.com/1261918
Backport-To: 3.6
Backport-To: 3.5
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/46005
Continuous-Integration: Jenkins CI
Reviewed-by: Michal Skrivanek <michal.skrivanek@redhat.com>
Reviewed-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>