Move jobs.py to lib/vdsm
The host jobs infrastructure in jobs.py is general purpose and belongs
in the vdsm-python package.  The qemuimg.py module (in vdsm-python)
wants to make use of jobs.py and we don't want to add a dependency from
qemuimg.py "library" to any python files in the vdsm package.

Change-Id: Icb8234d4a033ccacabfe292e6ec1baf3b2d2e18f
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/46141
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>