vm: sercon: serial config must match console's
When we add a serial console device, we also
inject a matching serial device.

Until now, the code assumed that the serial console has to be
bind to a PTY, while we can also bind it to an UNIX domain socket
(required by oVirt 3.6 "guest serial console" feature).

This patch makes sure that the injected serial device is bound
to the appropriate host endpoint, either pty or unix domain socket.
Please note that if the mismatch the serial device and the console
device configuration, the VM fails to start.

Change-Id: I49489a76517975fc40a226d2529bdd9f8d8c969c
Bug-Url: https://bugzilla.redhat.com/1223671
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/46187
Reviewed-by: Martin Polednik <mpolednik@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>