tests: Generalize sdm_indirection_tests
The infrastructure in sdm_indirection_tests is useful for testing the
other classes where there is indirection:

    Image -> ImageManifest
    Volume -> VolumeMetadata

In order to reuse this code we need to generalize two things:
1. The class doing the redirection is not necessarily a StorageDomain.
It can also be an Image or Volume.
2. Each class that redirects calls for SDM contains an instance of the
redirection target class.  For example, StorageDomains contain a field
'_manifest' which is a StorageDomainManifest instance.

The generalization can be done by extracting the general purpose bits
from DomainTestMixin into SDMIndirectionMixin.  When Image and Volume
support will be added to this module we can subclass
SDMIndirectionMixin as we are now doing for DomainTestMixin.

Change-Id: Icb68cb27d4e36c22f74a1cc445f38bea4d0ce4da
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/46259
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI