hsm: Add initialSize argument to createVolume
Add the possibility to specify the volume initial size,
for block disk with thin provisioning.

This argument is needed when importing a VM from an external
provider along with its disks. If one of the VM existing disk
is a block disk with thin provisioning, the new volume needs
a starting size as the original disk size.

If the initial size is not specified, the default initial
volume size for thin provisioning is the configured
volume_utilization_chunk_mb (1GiB by default)

Change-Id: Iaf5f142541bf0b311a77a0544a1c4ffb689a9fde
Bug-Url: https://bugzilla.redhat.com/1221603
Signed-off-by: Fred Rolland <frolland@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/46417
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Tested-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>
Continuous-Integration: Jenkins CI