tests: Fix and simplify cleanup after failures
If mount or umount failed, temporary images were not removed because of
incorrect cleanup code. This was not an issue since the temporary files
were created in a temporary directory, but we had two of them, so
failures to remove the first would leave also the second.

This patch simplify cleanup by removing the incorrect cleanup of the
temporary images, as they are removed anyway when the temporary
directory is removed.

Instead of two temporary directories, we use now one directory, and
create the two subdirectories inside it. Now we have to remove only one
directory in tearDown. To make it easier to debug, the temporary
directory is created with a prefix.

Change-Id: Iae829de5ae37008eff3154ad97d5f86c92a41eb2
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/46387
Continuous-Integration: Jenkins CI
Reviewed-by: Shmuel Leib Melamud <smelamud@redhat.com>
Reviewed-by: Sandro Bonazzola <sbonazzo@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>