tests: Fix unsafe monkeypatching
mkimage module could be left in bad state if setup failed after it was
patched, or removing temporary files in tearDown failed.

This patch ensures that the module is patched only if all other setUp
code was successful, and is restored first in tearDown.

The manual patching code was replaced with the monkeypatch.Patch.

Change-Id: I479fbd03db65e4c9cfa63fe4fe101f1b4b4c3876
X-Bug-Url: https://bugzilla.redhat.com/1268229
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/46388
Reviewed-by: Sandro Bonazzola <sbonazzo@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/47344
Tested-by: Francesco Romani <fromani@redhat.com>