netinfoTests: do not overwrite /etc/resolv.conf
os.path.join ignores any initial path components if it is passed
an absolute path, so a relative path must be used. Now, 'etc'
directory is never created and for test purposes it is not necessary
to do so, therefore stick with temp_dir and 'resolv.conf'.

Change-Id: I0985ce6e664d1cb50b63c6cbec7c8ca87899be18
Signed-off-by: Ondřej Svoboda <osvoboda@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/46396
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>