muiltipath: Unbreak configure on ovirt node
In commit commit 4d51f0df2d (multipath: Write multipath.conf atomically)
we replaced file copy with rename. This make the operation atomic but it
does not work on ovirt node, because multipath.conf is a bind mount.

On ovirt node we cannot do atomic rename; we must unpersist the
multipath.conf before the rename and persist it after the rename. If we
crash after unpersisting and before persisting multipath.conf will be
missing after a reboot. We can do nothing about this.

Other platforms are not effected since utils.unpersist() and
utils.persit() do nothing on other platforms.

Change-Id: I83822494dd1c98105aac081df0bdf4ec2e39a548
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/46397
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>