multipath: Fix multipath.conf selinux labels
When configuring multipath.conf, we were leaving multipath.conf with
incorrect selinux label:

    # ls -z /etc/multipath.conf
    -rw-r--r--. root root unconfined_u:object_r:etc_t:s0 /etc/multipath.conf

Now we restore the default selinux context on the temporary file before
renaming, so the final selinux labels are:

    # ls -z /etc/multipath.conf
    -rw-r--r--. root root system_u:object_r:etc_t:s0 /etc/multipath.conf

Due to a bug in libselinux python bindings [1], we set the label only if
selinux is enabled. This bug is fixed in upstream now.

[1] https://bugzilla.redhat.com/1264701

Change-Id: Id0aebae46f30abf908a8f48c4f1f2257394fe616
Backport-To: 3.6?
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/46398
Reviewed-by: Fabian Deutsch <fabiand@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>