virt: graphics: avoid None in specParams
In commit bbe8a03 we added a check to make sure
we use displayNetwork from global Vm configuration,
in the case the more specific setting was not sent
in the graphics device specParams.

One unfortunate side effect of this change is that
now None may end up in the graphics device specParams,
breaking XML-RPC communications.

We default (end recommend) JSON-RPC, but we still
support XML-RPC, and we use it as default in vdsClient,
so this must be avoided.

This patch reorganize the check to avoid the unwanted
injection of the None value.

Change-Id: I2a651c18c02c207f5497f22f0aad3d2967efd9a4
Backport-To: 3.6
Related-To: https://bugzilla.redhat.com/1261007
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/46629
Continuous-Integration: Jenkins CI
Reviewed-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>