gitignore: add tags
Ctags generate tag files for source code. Most of the approaches for
storing tags use the root directory of a project. This patch adds the
'tags' file to gitignore, causing git to ignore the tags file. This is
helpful when calling e.g. git clean, because the tags would be deleted.

Change-Id: Ic41dd126dd00af7d8f00cfbd21a9aecee15f9ea9
Signed-off-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/46664
Reviewed-by: Francesco Romani <fromani@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>