net: tests: test if setupNetworks drops/keeps initial nics IP
When we create a network over a nic, nics initial IP configuration
is dropped. When we create a bond over a nic, nics initial IP
configuration is kept.

This patch introduces two new tests which make sure about that and
could be used for test-driven developement of other configurators
and OVS hook.

Signed-off-by: Petr Horáček <phoracek@redhat.com>
Change-Id: I53068584d873b9349914262341e2a0a8cf4a4461
Reviewed-on: https://gerrit.ovirt.org/46690
Continuous-Integration: Jenkins CI
Reviewed-by: Ondřej Svoboda <osvoboda@redhat.com>
Reviewed-by: Ido Barkan <ibarkan@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>