vdsm-tool: Do not panic with sanlock empty groups
This fixes the following error:

    File "/usr/lib/python2.6/site-packages/vdsm/tool/configurators/
        sanlock.py", line 87, in isconfigured
    .strip().split(" ")]
    ValueError: invalid literal for int() with base 10: ''

Which happens when sanlock "Groups:" line is empty:

    $ grep "Groups:" /proc/<sanlock pid>/status
    Groups:

Sanlock must have its groups properly setup, but we should not
blow up because of this.

Fixed by using split() instead of split(' '), which cannot handle empty
string and is the wrong way to split in most cases.

Change-Id: I47f080beb748353970dec0753c9a0a7b1dd09bc6
Bug-Url:https://bugzilla.redhat.com/1267444
Backport-To: 3.6
Backport-To: 3.5
Signed-off-by: Germano Veit Michel <germano@redhat.com>
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/46742
Reviewed-by: Francesco Romani <fromani@redhat.com>
Continuous-Integration: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/47104
Continuous-Integration: Jenkins CI
Reviewed-by: Allon Mureinik <amureini@redhat.com>