netinfo: do not mistake a veth device for a VLAN in ifaceUsed
A heuristic in the function looks for a 'device' property of a network
link to decide if a given device "backs" a VLAN and thus is "used".
Not only VLANs have the property, however. At least veth devices have it
as well. This patch improves detection by exploiting the 'type' property
to match only 'vlan's.

At least testSetupNetworksAddDelDhcp was affected by this misdetection
on Fedora 22. The presence of 'device' property may be due to netlink
library but this is just a guess. (Also, iproute2 package/ip binary
recently started showing veth1@veth2 in its output but this need not
be related either.)

Change-Id: Iee3b45fb38f136bf69c175cbd745c89f80c299cf
Signed-off-by: Ondřej Svoboda <osvoboda@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/46867
Reviewed-by: Ido Barkan <ibarkan@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>