asyncore: use default handing of read event
It seems that we provided our own handle_read_event method due to
assumption that we need to distinguish between fds which support
pending and not. There are two fd types which do not provide pending
method: plain socket and AsyncoreEvent which uses EventFD. Only our
implementations of sslsocket provide pending.

Default implementation do not care about it because it is not using
pending and after several iteration of testing it seems to be OK.

Change-Id: Ia4dd6856f5c08da459478c60ac1fe7881baa69a2
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>
Reviewed-on: https://gerrit.ovirt.org/46911
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Yaniv Bronhaim
Reviewed-by: Dan Kenigsberg <danken@redhat.com>