hooks: ovs: move libvirt configuration into a separate module
Libvirt configuration has nothing to do with OVS configuration.

The main reason is to move creation of new libvirt networks to a place
where almost all configuration is completed, so an eventual rollback
doesn't have to do any extra work. This is necessary for the following
'improve ovs rollback' patch.

Change-Id: I775edc776c738822ed955ffb0b7885dae450b667
Signed-off-by: Petr Horáček <phoracek@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/46910
Reviewed-by: Ondřej Svoboda <osvoboda@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Ido Barkan <ibarkan@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>