stomp: reducing log footprint from stomp lib
There was bunch of not needed information about being not able to
process disconnect frame (part of the spec and expected). We made
vdsm aware of disconnect frame and few headers which are required
to validate the frame. 

We removed log statement which generate the most information
about message being sent but it do not bring any value when
debugging issues.


Bug-Url: https://bugzilla.redhat.com/1239062
Change-Id: I98f0c443fd50ec9e05dcf71ee1c7baabd044f48d
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>
Reviewed-on: https://gerrit.ovirt.org/46931
Continuous-Integration: Jenkins CI
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>