tests: fixed mangled names in crossImportTests
crossImportTests needs to extract the module name given
a python file. To do this, it uses a naive approach stripping the
extension of a filename, being the extension either ".py" or ".pyc".
This approach seems good enough for the test purposes, but it was
done incorrectly using the strip() method of string objects,
like

  filename.strip(".pyc")

Unfortunately, as per strip() documentation:

  strip(...)
    S.strip([chars]) -> string or unicode

    [...]
    If chars is given and not None, remove characters in chars instead.
    [...]

This means that *any* occurrence of *any* char provided will be removed,
not just the substring as a whole. This leads to incorrect output.
For example

  "cmdutils.pyc".strip(".pyc") -> "mdutils"

This patch fixes that using the more suitable os.path.splitext()
function. We take the chance to add an unit test for the get_mods()
helper, to make sure this won't happen again (or at least, not so
easily).

Change-Id: I53ba5fbc2f43fbafe9f6f291fcf2402986a5948c
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/46939
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>