schema: rpm for jsonrpc schema files
We need to use schema files outside of vdsm for external clients and in
the future for the engine. In order to do it we need to have separate
rpm which provides required files.

We introduced new package which is used for storing vdsmapi module. Spec
files are stored in the same places as before.

Change-Id: I13d6291ddbf5bf7d8e6a0956db3300cd0c45e563
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>
Reviewed-on: https://gerrit.ovirt.org/45750
Continuous-Integration: Jenkins CI
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>