net: Support VLAN hardware offloading in hostQOS
Apparently, the tc filter that VDSM uses to classify 802.1q traffic
is not working when the kernel offloads VLAN tagging to the underlying
physical nic.
This is because the VLAN field in the link layer header is only added
by the hardware and is not present by the time the filter inspects it.
Since the filter uses a straight forward matcher which only looks at
the headers it fails to match and the packets aren't classified to the
desired queuing discipline (qdisc).
Using a kernel packet meta data matcher should work regardless of VLAN
tagging offloading is enabled.

Change-Id: I667a9f38f4314da309685f6ba247f705a2e9c23e
Signed-off-by: Ido Barkan <ibarkan@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1271205
Reviewed-on: https://gerrit.ovirt.org/47443
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Jenkins CI