tests: Skip test if mounting loop device failed
Most of the mount loop device error are eliminated now, but
mkimage.mkIsoFs still fail randomly in the CI.

Since this is not a test failure but error in setting up the test, there
is no point in failing the build randomly, annoying developers working
on unrelated changes.

We will have to dive into this error later to understand why it happens
and what is the real fix to avoid it.

Change-Id: Ibb1041580f9e7b79d7d885c99f16e774daf7a6ec
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/47301
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Jenkins CI
Tested-by: Petr Horáček <phoracek@redhat.com>